Extract app into its own module
